refactored to move all the threading logic into a separate file
